spliting tests, fixing validators
