+++ setup.cfg
@@ -101,6 +101,11 @@ show_column_numbers = True
 show_error_codes = True
 ignore_missing_imports = True
 
+[coverage:run]
+omit =
+    npe2/manifest/configuration.py
+    npe2/manifest/keybindings.py
+
 [coverage:report]
 exclude_lines =
     pragma: no cover
@@ -110,3 +115,4 @@ exclude_lines =
     if __name__ == .__main__.:
     @overload
     @abstractmethod
+    except ImportError
