+++ tests/sample/my_plugin-1.2.3.dist-info/METADATA
@@ -1,3 +1,6 @@
 Metadata-Version: 2.1
 Name: my_plugin
 Version: 1.2.3
+Summary: My napari plugin
+License: BSD-3
+Author: The Black Knight
