+++ tests/test_cli.py
@@ -1,0 +1,13 @@
+import sys
+
+import pytest
+
+from npe2.cli import main
+
+
+def test_cli(monkeypatch, sample_path):
+    cmd = ["npe2", "validate", str(sample_path / "my_plugin" / "napari.yaml")]
+    monkeypatch.setattr(sys, "argv", cmd)
+    with pytest.raises(SystemExit) as e:
+        main()
+    assert e.value.code == 0
